rename `index_map` -> `index_meta`
