+++ hub/core/chunk_engine/write.py
@@ -1,5 +1,5 @@
 from hub.util.index import Index
-from hub.core.meta.index_map import IndexMeta
+from hub.core.meta.index_meta import IndexMeta
 from typing import List, Tuple
 from uuid import uuid1
